chore: rename use-toast file to useToast for consistency
