+++ commitment/imports/ui/components/base/toaster.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import { useToast } from "@hook/use-toast"
+import { useToast } from "@hook/useToast"
 import {
   Toast,
   ToastClose,
